Change directory path in build_services.sh
The script has been modified to change the directory path prior to the build and push operation in build_services.sh. This update helps to navigate directly to the petstore directory before iterating through petstore services to build, resulting in cleaner code and ease of debugging. The directory is reverted back to module-5 at the end.