Enable script executions and add App Insights setup
Uncommented lines in the setup script to enable the execution of `create_resources.sh` and `build_services.sh`. Renamed the environmental variable `RESOURCES_GROUP` from `rg-module5` to `module5_rg` for consistency reasons. A section was added at the end of the script to create an Application Insights resource in Azure, improving monitoring capabilities.