Add ContainerEnvironment class in PetStore app
This commit introduces a new class, `ContainerEnvironment`, in the PetStore application. This class serves as a Singleton to store the state of the container and perform other related tasks such as sending current users (via @Scheduled), generating JWT tokens, and initializing the environment with details from the version.json file and system properties.