Add header template to Pet Store App
A new header template is introduced to the Pet Store application. This includes options related to user authentication, such as profile editing, password reset, and logout. It also includes the user greeting section with Azure Dev Course mention and the shopping cart status. The header will be visible throughout all the pages of the application.