Add initial PetstoreappApplication class
A new main class named PetstoreappApplication was created as the entry point for the PetStore App. It includes a SpringBootApplication annotation and a main method which starts up the application and logs the startup. It also attempts to get the IP address of "www.google.com" and log it, throwing a RuntimeException if this fails.