Add products.html for Pet Store app
The new file, products.html, has been added to the pet store application. This improvement includes selectable products along with its image, name and a button to add to the cart. It also makes use of existing header, footer and disclaimer fragments, and implements a JavaScript function to manipulate the shopping cart.