Add ContainerEnvironment class in pet model
The new class, ContainerEnvironment, has been introduced to the pet model. It is designed to hold details about the container's environment, such as the host name, the application version and date, and the year. @PostConstruct is used for initialization tasks allowing the setting of these details during the creation of an instance.